crayonrobo/requirements.txt
click==8.1.7
cmake==3.28.1
comm==0.2.1
dash==2.17.0
ffmpy==0.3.1
ftfy==6.1.3
future==0.18.3
h5py==3.10.0
huggingface-hub==0.23.0
imageio==2.33.1
Jinja2==3.1.3
jupyter-client==7.3.4
matplotlib==3.7.4
matplotlib-inline==0.1.6
mpmath==1.3.0
multidict==6.0.5
networkx==3.1
numpy==1.24.4
oauthlib==3.2.2
open3d==0.18.0
openai==0.28.0
openai-clip==1.0.1
opencv-python==4.9.0.80
packaging==23.2
pandas==2.0.3
parso==0.8.3
pexpect==4.9.0
pickleshare==0.7.5
pillow==10.2.0
pip==24.0
pkgutil_resolve_name==1.3.10
platformdirs==4.2.0
plotly==5.22.0
prompt-toolkit==3.0.42
protobuf==4.25.2
PyYAML==6.0.1
requests==2.31.0
safetensors==0.4.2
sapien==0.7.0.dev0
scikit-learn==1.2.2
scipy==1.10.1
segment-anything==1.0
semantic-version==2.10.0
sentencepiece==0.1.99
sentry-sdk==1.42.0
service-identity==24.1.0
setproctitle==1.3.3
setuptools==68.2.2
shapely==2.0.3
shellingham==1.5.4
shortuuid==1.0.13
six==1.16.0
tensorboard==2.14.0
timm==0.6.13
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
tokenizers==0.13.3
torch==2.0.1
torchvision==0.15.2
tornado==6.1
tqdm==4.66.1
transformers==4.31.0
transforms3d==0.4.1
urllib3==2.1.0
wandb==0.16.4
wheel==0.41.2